lab_05/img-hidden/requirements.txt

Pillow
cryptography